requirements.txt
google-generativeai
aiohttp
sentence-transformers
numpy
scikit-learn
tenacity
reportlab
python-dotenv
pypdf
markdown
xhtml2pdf
networkx
pyvis
